Plot temporal de um parametro (com filtro por rio e/ou ponto)

Gera grafico temporal para um parametro, com opcoes de filtro por rios e/ou pontos. Se houver mais de um ponto, a cor diferencia pontos; opcional facet = TRUE para facetar por ponto. Pode adicionar reta de tendencia com add_trend = TRUE (lm).

Usage

param_plot(
  df,
  parametro,
  rios = NULL,
  pontos = NULL,
  add_trend = TRUE,
  facet = FALSE
)

Arguments

df

Data frame com data e a coluna do parametro. Idealmente contem ponto, e opcionalmente rio.

parametro

Character; nome do parametro.

rios

Vetor de nomes de rio a filtrar (opcional; usa coluna rio se existir).

pontos

Vetor de pontos a filtrar (opcional; usa coluna ponto se existir).

add_trend

Logical; se TRUE, adiciona geom_smooth(method = "lm", se = FALSE).

facet

Logical; se TRUE e houver ponto, aplica facet_wrap(~ponto).

Value

Objeto ggplot.

See also

Examples

if (FALSE) { # \dontrun{
data("wq_demo", package = "tikatuwq")
param_plot(wq_demo, "turbidez", pontos = c("P1","P2"), add_trend = TRUE, facet = TRUE)
} # }