Resumo estatistico por parametro (com filtro por rio e/ou ponto)

Produz resumo estatistico para um parametro, com opcoes de filtro por rios e/ou pontos e agregacao opcional por periodo (mes/trimestre/ano), quando houver coluna data.

Usage

param_summary(
  df,
  parametro,
  rios = NULL,
  pontos = NULL,
  period = c("none", "month", "quarter", "year"),
  na_rm = TRUE
)

Arguments

df

Data frame com ao menos a coluna do parametro. Idealmente contem ponto, e opcionalmente rio e data.

parametro

Character; nome do parametro (ex.: "turbidez", "od", "pH").

rios

Vetor de nomes de rio a filtrar (opcional; usa coluna rio se existir).

pontos

Vetor de pontos a filtrar (opcional; usa coluna ponto se existir).

period

"none", "month", "quarter" ou "year" para agregar por periodo (requer coluna data).

na_rm

Remover NA dos calculos? (default TRUE)

Value

Tibble com colunas de agrupamento disponiveis (rio, ponto, periodo) e metricas: n, mean, sd, min, median, max.

See also

Examples

if (FALSE) { # \dontrun{
data("wq_demo", package = "tikatuwq")
param_summary(wq_demo, "turbidez", pontos = "P1")
param_summary(wq_demo, "od", rios = "Rio Azul", period = "month")
} # }