Time series by parameter

Plot a time series for one numeric parameter, optionally colored/faceted by a grouping column.

Usage

plot_series(df, parametro, facet = NULL)

Arguments

df

Data frame with a data column (Date/POSIXct) and the parameter column.

parametro

Character; name of the numeric column to plot on Y.

facet

Character or NULL; optional grouping column name to color/facet.

Value

A ggplot object.

See also

Examples

# \donttest{
data(wq_demo)
#> Warning: data set 'wq_demo' not found
# Basic: time series of turbidity
p <- plot_series(wq_demo, "turbidez")
#> Error: object 'wq_demo' not found
# With color/facet by sampling point
p2 <- plot_series(wq_demo, "turbidez", facet = "ponto")
#> Error: object 'wq_demo' not found
# }